chore: migrate to vite and vitest (#200)
Migrates build and tests to vite. The motivation is to make it a bit
easier to support React and JSX in a future where we need something for
MCP-UI or ChatGPT apps. We may also make use of [vite's importing assets
as strings](https://vite.dev/guide/assets.html#importing-asset-as-string)
for HTML or CSS templates.

Most of the changes are in the tests, that change a bit from jest to
vitest.