feat!: semantic priority values (P1, P2, P3, P4) instead of numeric (#217)
> [!WARNING]
> This is a breaking change.

## 🗺️ Overview

Two things:

- What the title says. Now our tools all return tasks data structures
where the priority is always expressed as p1, p2, p3 or p4, and never as
a number.
- In a 2nd commit there's an attempt to promote not using numeric
literals for priority in the code, even in place where the number value
is expected. So instead of saying `priority: 1` we'll now say `priority:
convertPriorityToNumber('p4')` instead.

See
https://github.com/Doist/todoist-ai/pull/216#pullrequestreview-3460766807
and the subsequent few comments below it.

## ✅ PR Checklist

-   [x] Added tests for bugs / new features

## 🎥 Demo


https://github.com/user-attachments/assets/69401b79-cf08-427f-b752-d6a8ade96995