feat: Add support for MCP Apps (task list resource) (#245)
## Short description

This PR adds basic scaffolding for [ChatGPT
Apps](https://openai.com/index/introducing-apps-in-chatgpt/). It also
adds the first app (or rather resource) for `find-tasks-by-date`. This
allows users to ask "show me my tasks for today" and instead of getting
a textual representation, they see a proper task list.

* Code Walkthrough: https://share.cleanshot.com/p102ckfB
* Local testing (Doister-only): https://share.cleanshot.com/5T3twn1H
* Demo in ChatGPT: https://share.cleanshot.com/JR5bWRQ4

---

We plan to further improve this first app by showing more task
attributes. This follow in a separate PR to keep scope in check:
- description and due date (simple)
- project and section (needs tool call)

---------

Co-authored-by: Henning Muszynski <henningmu@users.noreply.github.com>
Co-authored-by: Ernesto García <ernesto@doist.com>